create distinct factory methods for each cluster type
