+++ docs/getting_started.rst
@@ -53,7 +53,7 @@ terminated) for you.
             )
 
     # Using a Cloud provider
-    cluster = rh.cluster(
+    cluster = rh.autocluster(
                 name="cpu-cluster",
                 instance_type="CPU:8",
                 provider="cheapest",      # options: "AWS", "GCP", "Azure", "Lambda", or "cheapest"
