+++ runhouse/__init__.py
@@ -18,7 +18,7 @@ from .rns.defaults import Defaults
 from .rns.envs import conda_env, CondaEnv, env, Env
 from .rns.folders.folder import folder, Folder
 from .rns.function import function, Function
-from .rns.hardware import cluster, Cluster, OnDemandCluster
+from .rns.hardware import autocluster, cluster, Cluster, OnDemandCluster
 from .rns.kvstores.kvstore import KVStore
 from .rns.login import login, logout
 from .rns.packages import git_package, GitPackage, package, Package
