+++ runhouse/rns/hardware/on_demand_cluster.py
@@ -2,7 +2,7 @@ import contextlib
 import logging
 import subprocess
 from pathlib import Path
-from typing import Any, Dict
+from typing import Any, Dict, Optional
 
 import ray
 
@@ -382,3 +382,83 @@ class OnDemandCluster(Cluster):
     def ssh(self):
         """SSH into the cluster."""
         subprocess.run(["ssh", f"{self.name}"])
+
+
+# OnDemandCluster factory method
+def autocluster(
+    name: str,
+    instance_type: Optional[str] = None,
+    num_instances: Optional[int] = None,
+    provider: Optional[str] = None,
+    autostop_mins: Optional[int] = None,
+    use_spot: bool = False,
+    image_id: Optional[str] = None,
+    region: Optional[str] = None,
+    dryrun: bool = False,
+) -> OnDemandCluster:
+    """
+    Builds an instance of :class:`OnDemandCluster`.
+
+    Args:
+        name (str): Name for the cluster, to re-use later on.
+        instance_type (int, optional): Type of cloud instance to use for the cluster. This could
+            be a Runhouse built-in type, or your choice of instance type.
+        num_instances (int, optional): Number of instances to use for the cluster.
+        provider (str, optional): Cloud provider to use for the cluster.
+        autostop_mins (int, optional): Number of minutes to keep the cluster up after inactivity,
+            or ``-1`` to keep cluster up indefinitely.
+        use_spot (bool, optional): Whether or not to use spot instance.
+        image_id (str, optional): Custom image ID for the cluster.
+        region (str, optional): The region to use for the cluster.
+        dryrun (bool): Whether to create the Cluster if it doesn't exist, or load a Cluster object as a dryrun.
+            (Default: ``False``)
+
+    Returns:
+        OnDemandCluster: The resulting cluster.
+
+    Example:
+        >>> # On-Demand SkyPilot Cluster (OnDemandCluster)
+        >>> gpu = rh.autocluster(name='rh-4-a100s',
+        >>>                  instance_type='A100:4',
+        >>>                  provider='gcp',
+        >>>                  autostop_mins=-1,
+        >>>                  use_spot=True,
+        >>>                  image_id='my_ami_string',
+        >>>                  region='us-east-1',
+        >>>                  )
+    """
+    if (
+        instance_type is None
+        and num_instances is None
+        and provider is None
+        and autostop_mins is None
+        and not use_spot
+        and image_id is None
+        and region is None
+        and dryrun
+    ):
+        # If only the name is provided and dryrun is set to True
+        return OnDemandCluster.from_name(name)
+
+    config = rns_client.load_config(name)
+    config["name"] = name or config.get("rns_address", None) or config.get("name")
+
+    config["instance_type"] = instance_type or config.get("instance_type", None)
+    config["num_instances"] = num_instances or config.get("num_instances", None)
+    config["provider"] = provider or config.get("provider", None)
+    config["autostop_mins"] = (
+        autostop_mins
+        if autostop_mins is not None
+        else config.get("autostop_mins", None)
+    )
+    config["use_spot"] = (
+        use_spot if use_spot is not None else config.get("use_spot", None)
+    )
+    config["image_id"] = (
+        image_id if image_id is not None else config.get("image_id", None)
+    )
+    config["region"] = region if region is not None else config.get("region", None)
+
+    new_cluster = OnDemandCluster.from_config(config, dryrun=dryrun)
+
+    return new_cluster
